Add files via upload
Made some changes because I was not receiving the entire speech-to-text string as expected.  The buffer size is currently set to 256, which was plenty large to hold a few words.  Next task, to strip the first two characters from the packetbuffer, and assign "inputs" as the rest of the  remainder of packetbuffer.  Also, this project will probably migrate to the Arduino Mega2560.  I will continue to make sure that the core functions will remain working with the Arduino Uno.