For those with an Arduino Mega
Use these files on your mega.  Note the pin changes, we are also using hardware serial (much, much, much, much better results when using the app).  Wire accordingly.  I'm using serial3 right now, any of the hardware serial ports should do, but this code is written for serial3.  So, Tx on the bluefruit to rx3, Rx on the bluefruit to tx3, cts to grnd, and grnd to grnd, and i set mode pin to digital 50.  Pinout change for mode pin is set in BluefruitConfig.h, and uncomment out any of the other spi/software serial for the bluefruit.